feat: add countries to store
