--- src/app/redux/slices/first.enum.ts
@@ -1,3 +1,0 @@
-export const enum Country {
-  COUNTRY_01 = 'COUNTRY_01',
-}
