--- src/app/redux/slices/first.type.ts
@@ -1,12 +1,0 @@
-import { Country } from './first.enum'
-
-export type SubmitData = {
-  name: string | null
-  age: number | null
-  email: string | null
-  password: string | null
-  gender: string | null
-  terms: boolean | null
-  picture: string | null //base64
-  country: Country | null
-}
