+++ src/app/redux/store.ts
@@ -1,9 +1,11 @@
 import { configureStore } from '@reduxjs/toolkit'
+import { countriesReducer } from './slices/countries.slice'
 import { firstReducer } from './slices/first.slice'
 
 export const store = configureStore({
   reducer: {
     firstForm: firstReducer,
+    countries: countriesReducer,
   },
 })
