+++ src/shared/lib/file-converter.ts
@@ -1,9 +1,10 @@
-export const fileToBase64 = (
-  file: File
-): Promise<string | ArrayBuffer | null> => {
+export const fileToBase64 = (file: File): Promise<string | null> => {
   const fr = new FileReader()
   return new Promise((res, rej) => {
-    fr.onloadend = (): void => res(fr.result)
+    fr.onloadend = (): void => {
+      const { result } = fr
+      return res(typeof result === 'string' ? result : null)
+    }
     fr.onerror = rej
     fr.readAsDataURL(file)
   })
